+++ tests/apps/courses/test_templates_course_detail_rendering.py
@@ -1267,3 +1267,108 @@ class RunsCourseCMSTestCase(CMSTestCase):
             '<li class="is-hidden">',
         )
         self.assertNotContains(response, "course-detail__view-more-runs")
+
+    def test_templates_course_detail_meta_description(self):
+        """
+        The course meta description should show meta_description placeholder if defined
+        """
+        course = CourseFactory()
+        page = course.extended_object
+
+        title_obj = page.get_title_obj(language="en")
+        title_obj.meta_description = "A custom description of the course"
+        title_obj.save()
+
+        page.publish("en")
+
+        url = course.extended_object.get_absolute_url()
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+
+        self.assertContains(
+            response,
+            '<meta name="description" content="A custom description of the course" />',
+        )
+
+    def test_templates_course_detail_meta_description_course_introduction(self):
+        """
+        The course meta description should show the course_introduction if no meta_description is
+        specified
+        """
+        course = CourseFactory()
+        page = course.extended_object
+
+        # Add an course_introduction to the course
+        placeholder = course.extended_object.placeholders.get(
+            slot="course_introduction"
+        )
+        add_plugin(
+            language="en",
+            placeholder=placeholder,
+            plugin_type="PlainTextPlugin",
+            body="A further course introduction of the course",
+        )
+        page.publish("en")
+
+        url = course.extended_object.get_absolute_url()
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+
+        self.assertContains(
+            response,
+            '<meta name="description" content="A further course introduction of the course" />',
+        )
+
+    def test_templates_course_detail_meta_description_course_introduction_max_length(
+        self,
+    ):
+        """
+        The course meta description should be cut if it exceeds more than 160 caracters
+        """
+        course = CourseFactory()
+        page = course.extended_object
+        placeholder_value = (
+            "Long description that describes the page with a summary. "
+            "Long description that describes the page with a summary. "
+            "Long description that describes the page with a summary. "
+        )
+
+        # Add an course_introduction to the course
+        placeholder = course.extended_object.placeholders.get(
+            slot="course_introduction"
+        )
+        add_plugin(
+            language="en",
+            placeholder=placeholder,
+            plugin_type="PlainTextPlugin",
+            body=placeholder_value,
+        )
+        page.publish("en")
+
+        url = course.extended_object.get_absolute_url()
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+
+        cut = placeholder_value[0:160]
+        self.assertContains(
+            response,
+            f'<meta name="description" content="{cut}" />',
+        )
+
+    def test_templates_course_detail_meta_description_empty(self):
+        """
+        The course meta description should not be present if neither the meta_description field
+        on the page, nor the `course_introduction` placeholder are filled
+        """
+        course = CourseFactory()
+        page = course.extended_object
+        page.publish("en")
+
+        url = course.extended_object.get_absolute_url()
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+
+        self.assertNotContains(
+            response,
+            '<meta name="description"',
+        )